pyproject.toml
[project]
name = "dist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]
